refactor: reorganize source files into utils and handlers, update import paths accordingly
