--- src/handlers/index.ts
@@ -1,4 +1,0 @@
-export { createConvertHandler } from './convertHandler';
-export { createPasteHandler } from './pasteHandler';
-export { createUndoHandler } from './undoHandler';
-export { createDecorationHandler } from './decorationHandler';
