+++ src/providers/I18nJumpProvider.ts
@@ -1,16 +1,16 @@
-import { EventEmitter, Range, workspace, Position, Uri } from 'vscode';
+import { Range, workspace, Position, Uri } from 'vscode';
 import { isNil, max } from 'lodash';
 import { match } from 'minimatch';
 import { AhoCorasick } from '@monyone/aho-corasick';
 
-import Hook from './hook';
-import I18n from './i18n';
-import { getConfig } from './config';
-import { checkSupportType } from './utils';
-import { MatchType, SupportType } from './constant';
+import Hook from '@/utils/hook';
+import I18n from '@/utils/i18n';
+import { getConfig } from '@/utils/config';
+import { MatchType, SupportType } from '@/utils/constant';
+import { checkSupportType } from '@/utils';
 
-import type { TextDocumentContentProvider, DefinitionProvider, TextDocument } from 'vscode';
-import type { I18nGroup } from './types';
+import type { DefinitionProvider, TextDocument } from 'vscode';
+import type { I18nGroup } from '@/types';
 
 const genLocationLink = ({ supportType = SupportType.All, range, locationLink, filePath, line = 0 }: I18nGroup) => {
     if (!checkSupportType(SupportType.Jump, supportType) || isNil(range)) {
@@ -31,7 +31,7 @@ const genLocationLink = ({ supportType = SupportType.All, range, locationLink, f
         targetRange: new Range(lineEndPos, lineEndPos),
         originSelectionRange: range,
     }];
-}
+};
 
 export class I18nJumpProvider implements DefinitionProvider {
     static instance: I18nJumpProvider;
@@ -68,47 +68,5 @@ export class I18nJumpProvider implements DefinitionProvider {
         }
 
         return genLocationLink((await Hook.getInstance().matchI18n({ type: MatchType.Document, i18nGroups: [{ ...group, range: matched.range }], document }))[0]);
-    }
-}
-
-export class MemoryDocumentProvider implements TextDocumentContentProvider {
-    private static instance: MemoryDocumentProvider;
-
-    static getInstance() {
-        if (!MemoryDocumentProvider.instance) {
-            MemoryDocumentProvider.instance = new MemoryDocumentProvider();
-        }
-        
-        return MemoryDocumentProvider.instance;
-    }
-
-    static async getDocument(documentText: string, uri?: Uri) {
-        const memoryDocumentProvider = MemoryDocumentProvider.getInstance();
-        const documentUri = uri || Uri.parse('memory://temp-document');
-        memoryDocumentProvider.updateDocument(documentUri, documentText);
-        const document = await workspace.openTextDocument(documentUri);
-
-        return document;
-    }
-
-    private _onDidChange = new EventEmitter<Uri>();
-    private documents: Map<string, string> = new Map();
-
-    get onDidChange() {
-        return this._onDidChange.event;
-    }
-
-    provideTextDocumentContent(uri: Uri) {
-        return this.documents.get(uri.toString()) || '';
-    }
-
-    updateDocument(uri: Uri, newContent: string) {
-        this.documents.set(uri.toString(), newContent);
-        this._onDidChange.fire(uri);
-    }
-
-    deleteDocument(uri: Uri) {
-        this.documents.delete(uri.toString());
-        this._onDidChange.fire(uri);
     }
 }
