+++ src/utils/config.ts
@@ -1,10 +1,9 @@
 import { workspace } from 'vscode';
 
-import { PLUGIN_NAME } from "./constant";
-import { ConflictPolicy } from './constant';
+import { PLUGIN_NAME, ConflictPolicy } from "@/utils/constant";
 
 import type { WorkspaceConfiguration } from "vscode";
-import type { Config } from './types';
+import type { Config } from '@/types';
 
 const defaultConfig: Config = {
     hookFilePattern: '.vscode/i18n-fast.hook.js',
@@ -15,8 +14,8 @@ const defaultConfig: Config = {
 
 const genConfig = (config: WorkspaceConfiguration): Config => {
     return Object.entries(defaultConfig).reduce((pre, [key]) => ({ ...pre, [key]: config.get(key) }), defaultConfig);
-}
+};
 
 export const getConfig = (): Config => {
     return genConfig(workspace.getConfiguration(PLUGIN_NAME));
-}
+};
