+++ tsconfig.json
@@ -7,6 +7,10 @@
 		],
 		"sourceMap": true,
 		"rootDir": "src",
+		"baseUrl": "./",
+		"paths": {
+			"@/*": ["src/*"]
+		},
 		"strict": true   /* enable all strict type-checking options */
 		/* Additional Checks */
 		// "noImplicitReturns": true, /* Report error when not all code paths in function return a value. */
